requirements.txt
numpy==2.2.1
opencv_python==4.10.0.84
Pillow==11.1.0
